chore: Export subtask argument types
