+++ packages/ua-devtools-evm-hardhat/src/index.ts
@@ -1,3 +1,4 @@
 export * from './constants'
 export * from './oapp'
 export * from './ownable'
+export * from './tasks/types'
