+++ packages/ua-devtools-evm-hardhat/src/tasks/oapp/types.ts
@@ -1,0 +1,10 @@
+import type { OmniGraphHardhat } from '@layerzerolabs/devtools-evm-hardhat'
+import type { ZodType, ZodTypeDef } from 'zod/lib/types'
+
+export * from './wire/types'
+
+export interface SubtaskLoadConfigTaskArgs {
+    configPath: string
+    schema: ZodType<OmniGraphHardhat, ZodTypeDef, OmniGraphHardhat>
+    task: string
+}
