+++ packages/ua-devtools-evm-hardhat/src/tasks/oapp/wire/types.ts
@@ -1,0 +1,7 @@
+import type { OAppOmniGraph, OAppConfigurator, OAppFactory } from '@layerzerolabs/ua-devtools'
+
+export interface SubtaskConfigureTaskArgs {
+    graph: OAppOmniGraph
+    configurator?: OAppConfigurator
+    oappFactory?: OAppFactory
+}
